chore : @NPlusOneTest에 대한 통합 테스트 환경의 검증을 표현하도록 클래스명 수정
